+++ hadoop-hdds/server-scm/src/test/java/org/apache/hadoop/hdds/scm/ha/TestReplicationAnnotation.java
@@ -16,27 +16,112 @@
  */
 package org.apache.hadoop.hdds.scm.ha;
 
+import org.apache.hadoop.hdds.protocol.proto.HddsProtos;
 import org.apache.hadoop.hdds.protocol.proto.SCMRatisProtocol.RequestType;
-import org.apache.hadoop.hdds.scm.metadata.Replicate;
+import org.apache.hadoop.hdds.scm.container.ContainerID;
+import org.apache.hadoop.hdds.scm.container.ContainerInfo;
+import org.apache.hadoop.hdds.scm.container.ContainerReplica;
+import org.apache.hadoop.hdds.scm.container.ContainerStateManagerV2;
+import org.apache.hadoop.hdds.scm.pipeline.PipelineID;
+import org.apache.hadoop.ozone.common.statemachine.InvalidStateTransitionException;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+
+import java.io.IOException;
+import java.util.Map;
+import java.util.NavigableSet;
+import java.util.Set;
 
 /**
  * Tests on {@link org.apache.hadoop.hdds.scm.metadata.Replicate}.
  */
 public class TestReplicationAnnotation {
   private SCMHAInvocationHandler scmhaInvocationHandler;
+  private ContainerStateManagerV2 mockCSM;
 
   @Before
   public void setup() {
     SCMHAManager mock = MockSCMHAManager.getInstance(true);
-    TestReplicationAnnotation testReplicationAnnotation =
-        new TestReplicationAnnotation();
+    mockCSM = new ContainerStateManagerV2() {
+      @Override
+      public boolean contains(HddsProtos.ContainerID containerID) {
+        return false;
+      }
+
+      @Override
+      public Set<ContainerID> getContainerIDs() {
+        return null;
+      }
+
+      @Override
+      public Set<ContainerID> getContainerIDs(
+          HddsProtos.LifeCycleState state) {
+        return null;
+      }
+
+      @Override
+      public ContainerInfo getContainer(HddsProtos.ContainerID id) {
+        return null;
+      }
+
+      @Override
+      public Set<ContainerReplica> getContainerReplicas(
+          HddsProtos.ContainerID id) {
+        return null;
+      }
+
+      @Override
+      public void updateContainerReplica(
+          HddsProtos.ContainerID id, ContainerReplica replica) {
+
+      }
+
+      @Override
+      public void removeContainerReplica(
+          HddsProtos.ContainerID id, ContainerReplica replica) {
+
+      }
+
+      @Override
+      public void addContainer(HddsProtos.ContainerInfoProto containerInfo)
+          throws IOException {
+
+      }
+
+      @Override
+      public void updateContainerState(
+          HddsProtos.ContainerID id, HddsProtos.LifeCycleEvent event)
+          throws IOException, InvalidStateTransitionException {
+
+      }
+
+      @Override
+      public void updateDeleteTransactionId(
+          Map<ContainerID, Long> deleteTransactionMap) throws IOException {
+
+      }
+
+      @Override
+      public ContainerInfo getMatchingContainer(
+          long size, String owner, PipelineID pipelineID,
+          NavigableSet<ContainerID> containerIDs) {
+        return null;
+      }
+
+      @Override
+      public void removeContainer(HddsProtos.ContainerID containerInfo)
+          throws IOException {
+      }
+
+      @Override
+      public void close() throws IOException {
+
+      }
+    };
+
     scmhaInvocationHandler = new SCMHAInvocationHandler(
-        RequestType.CONTAINER,
-        testReplicationAnnotation,
-        mock.getRatisServer());
+        RequestType.CONTAINER, mockCSM, mock.getRatisServer());
   }
 
   @Test
@@ -44,12 +129,15 @@ public class TestReplicationAnnotation {
     // test whether replicatedOperation will hit the Ratis based replication
     // code path in SCMHAInvocationHandler. The invoke() can return means the
     // request is handled properly thus response is successful. Expected
-    // result is null because replicatedOperation returns nothing.
+    // result is null because the funciton returns nothing.
+    Object[] arguments = {HddsProtos.ContainerInfoProto.getDefaultInstance()};
     Assert.assertEquals(null, scmhaInvocationHandler.invoke(new Object(),
-        this.getClass().getMethod("replicatedOperation"), new Object[0]));
+        mockCSM.getClass().getMethod(
+            "addContainer", HddsProtos.ContainerInfoProto.class),
+        arguments));
   }
 
-  @Replicate
-  public void replicatedOperation() {
-  }
+//  @Replicate
+//  public void replicatedOperation() {
+//  }
 }
